requirements.txt
aiohttp==3.6.2
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
fastapi==0.55.1
h11==0.9.0
hashids==1.2.0
idna==2.9
idna-ssl==1.1.0
multidict==4.7.6
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-engineio==3.14.2
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
starlette==0.13.2
typing-extensions==3.7.4.2
uvicorn==0.11.7
websockets==8.1
yarl==1.4.2
cryptography==3.3.1